+++ src/main.rs
@@ -3,6 +3,7 @@ mod cas_err;
 mod postgres;
 use crate::cas_err::CasErr;
 use postgres::connection::{Conn, ConnectionParams};
+mod cas_val;
 mod json;
 
 fn main() {
@@ -26,6 +27,10 @@ fn exec_query() -> Result<(), CasErr> {
     };
     let mut conn = Conn::connect(params)?;
     // conn.query(String::from("SELECT * FROM pg_type"), vec![], json::write_json)
-    conn.query(String::from("SELECT typname FROM pg_type LIMIT 10"), vec![], json::write_json)
+    conn.query(
+        String::from("SELECT typname FROM pg_type LIMIT 10"),
+        vec![],
+        json::write_json,
+    )
     // conn.query(String::from("SELECT * FROM points"), vec![])
 }
